Deleted unused components
